pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobscraper"
version = "0.1.0"
description = "Apify-powered job scraping and OpenAI job-fit evaluation pipeline."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1,<4",
    "openai>=1.93,<3",
    "requests>=2.31,<3",
    "tzdata>=2024.1",
    "google-api-python-client>=2.100,<3",
    "google-auth>=2.20,<3",
    "google-auth-httplib2>=0.2,<1",
    "google-auth-oauthlib>=1.0,<2",
]

[project.scripts]
jobscraper-scrape = "jobscraper.scraper.cli:main"
jobscraper-evaluate = "jobscraper.evaluator.cli:main"
jobscraper-pipeline = "jobscraper.pipeline.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
